requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
orjson>=3.8.0
requests>=2.25.0
click>=8.0.0

# Embedding providers
sentence-transformers>=2.2.0
google-cloud-aiplatform
accelerate

# Tree-sitter for semantic code chunking
tree-sitter>=0.25.0

# Cloud storage support
boto3>=1.26.0  # For S3 storage backend
google-cloud-storage

# Web API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Optional: For development and testing
# pytest>=7.0.0
# flake8>=6.0.0
